\name{GLS-package}
\alias{GLS}
\docType{package}
\title{
  GLS - scanGLS
}
\description{
  TODO: Description of the package
}
\details{
  TODO: Links to the main functions of the package
  \itemize{
    \item \code{\link{scanGLS}} - \code{\link{scanGLS}}
  }
}
\author{
  Willem Kruijer \email{willem.kruijer@wur.nl}\cr
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Willem Kruijer \email{willem.kruijer@wur.nl}\cr
  Contributions from: ...
}
\references{
  \itemize{
    \item TODO
  }
}
\keyword{package}

\seealso{
  \itemize{
    \item \code{\link{scanGLS}} - main function in scanGLS
  }
}
