\name{scan.GLS}
\alias{scan.GLS}
\title{scan.GLS - GWAS analysis }
\description{
  GWAS analysis
}
\usage{
scan.GLS(gwas.obj, input.pheno, varcomp.file, output.file, covariate.file="")
}

\arguments{
  \item{gwas.obj}{ TODO }
  \item{input.pheno}{ TODO }
  \item{varcomp.file}{ TODO }
  \item{output.file}{ TODO }
  \item{covariate.file}{ TODO }
}
\details{
  TODO
  \itemize{
    \item TODO
  }
}
\value{
  TODO
}
\references{ 
  TODO
}
\author{
  Willem Kruijer \email{willem.kruijer@wur.nl}\cr
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Willem Kruijer \email{willem.kruijer@wur.nl}\cr
  Contributions from: ...
}
\note{
  TODO
}
\examples{
#TODO
}

\keyword{methods}
